refactor: reduce book updates
- Move the (very similar) reduce functions of invoice and orders into a shared location, including all functionality.
- Existing tests should cover the refactor.
